requirements.txt
# Python 3.10
pillow>=9.5.0
scikit-image>=0.21.0
onnxruntime>=1.15.0
numpy
tqdm
